DLL
Added DLL for setting the players positions in both teams/ this has been implemented in the DLLTEST c sharp file in scripts. Its on MIke scene on the sky box for testing. If you want to change the size you will need to do so from the actuall dll. it has also implemented in the spawnpoint class but we have not tried it as we can not start the game due to UI being incomplete. You can fidn the dll In Game-Optimization-Project/SpawnPointsDuckHuntDLL